Reflected type hint to be float

+++ displaylib/pygame/constants.py
@@ -2,6 +2,6 @@ from __future__ import annotations
 
 
 DEFAULT: int = 0
-MILLISECOND: int = 1000.0
+MILLISECOND: float = 1000.0
 
 __all__ = ["DEFAULT", "MILLISECOND"]
